+++ rwrs.py
@@ -1,12 +1,9 @@
-from flask import Flask, render_template, make_response, abort, request, redirect, url_for, flash, g
+from flask import Flask, render_template, make_response, request, g
 from werkzeug.exceptions import HTTPException
 from flask_sqlalchemy import SQLAlchemy
-from sqlalchemy_utils import ArrowType
 from flask_caching import Cache
 from helpers import *
 import logging
-import arrow
-import click
 import math
 import sys
 import os
@@ -70,273 +67,9 @@ for handler in app.logger.handlers:
 
 
 import rwr
-
-
-# -----------------------------------------------------------
-# Routes
-
-
-@app.route('/')
-def home():
-    return render_template(
-        'home.html'
-    )
-
-
-@app.route('/my-friends')
-def my_friends():
-    return render_template(
-        'manage_friends.html'
-    )
-
-
-@app.route('/players')
-@app.route('/players/<username>')
-def player_stats(username=None):
-    if not username:
-        username = request.args.get('username').strip()
-
-        # Redirect to a SEO-friendly URL if the username query parameter is detected
-        return redirect(url_for('player_stats', username=username))
-
-    if not username:
-        abort(404)
-
-    scraper = rwr.DataScraper()
-
-    player = scraper.search_player(username)
-
-    if not player:
-        flash('Sorry, the player "{}" wasn\'t found. Maybe this player hasn\'t already played on a ranked server yet.'.format(username), 'error')
-
-        return redirect(url_for('home'))
-
-    servers = scraper.get_servers()
-
-    player.set_playing_on_server(servers)
-
-    return render_template(
-        'player_stats.html',
-        player=player
-    )
-
-
-@app.route('/players/<username>/compare')
-@app.route('/players/<username>/compare/<username_to_compare_with>')
-def players_compare(username, username_to_compare_with=None):
-    if not username_to_compare_with:
-        username_to_compare_with = request.args.get('username_to_compare_with').strip()
-
-        # Redirect to a SEO-friendly URL if the username_to_compare_with query parameter is detected
-        return redirect(url_for('players_compare', username=username, username_to_compare_with=username_to_compare_with))
-
-    if not username_to_compare_with:
-        abort(404)
-
-    scraper = rwr.DataScraper()
-
-    player = scraper.search_player(username)
-
-    if not player:
-        flash('Sorry, the player "{}" wasn\'t found. Maybe this player hasn\'t already played on a ranked server yet.'.format(username), 'error')
-
-        return redirect(url_for('home'))
-
-    player_to_compare_with = scraper.search_player(username_to_compare_with)
-
-    if not player_to_compare_with:
-        flash('Sorry, the player "{}" wasn\'t found. Maybe this player hasn\'t already played on a ranked server yet.'.format(username_to_compare_with), 'error')
-
-        return redirect(url_for('player_stats', username=username))
-
-    servers = scraper.get_servers()
-
-    player.set_playing_on_server(servers)
-
-    return render_template(
-        'player_stats.html',
-        player=player,
-        player_to_compare_with=player_to_compare_with
-    )
-
-
-@app.route('/servers')
-def servers_list():
-    scraper = rwr.DataScraper()
-
-    filters = request.args.to_dict()
-
-    if filters:
-        servers = scraper.filter_servers(**filters)
-    else:
-        servers = scraper.get_servers()
-
-    locations = scraper.get_all_servers_locations()
-    types = scraper.get_all_servers_types()
-    modes = scraper.get_all_servers_modes()
-    maps = scraper.get_all_servers_maps()
-
-    return render_template(
-        'servers_list.html',
-        servers=servers,
-        locations=locations,
-        types=types,
-        modes=modes,
-        maps=maps
-    )
-
-
-@app.route('/servers/<ip_and_port>')
-def server_details(ip_and_port):
-    ip, port = ip_and_port.split(':', maxsplit=1)
-
-    scraper = rwr.DataScraper()
-
-    server = scraper.search_server(ip, int(port))
-
-    if not server:
-        flash('Sorry, this server wasn\'t found.', 'error')
-
-        return redirect(url_for('servers_list'))
-
-    return render_template(
-        'server_details.html',
-        server=server
-    )
-
-
-# -----------------------------------------------------------
-# Models
-
-
-class ServerPlayerCount(db.Model):
-    TIMESPAN_LAST_DAY = 1
-    TIMESPAN_LAST_WEEK = 2
-    TIMESPAN_LAST_MONTH = 3
-
-    class ServerPlayerCountQuery(db.Query):
-        def _get_base_count_query(self, timespan):
-            pass # TODO
-
-        def get_player_count(self, timespan, ip=None, port=None):
-            pass # TODO
-
-        def get_server_count(self, timespan, online_only=False):
-            pass # TODO
-
-    __tablename__ = 'servers_player_count'
-    __bind_key__ = 'servers_player_count'
-    __table_args__ = (db.Index('ip_port_idx', 'ip', 'port'), )
-    query_class = ServerPlayerCountQuery
-
-    id = db.Column(db.Integer, primary_key=True, autoincrement=True) # TODO To remove because useless
-
-    _ip = db.Column('ip', db.Integer, nullable=False)
-    port = db.Column(db.Integer, nullable=False)
-    measured_at = db.Column(ArrowType, default=arrow.utcnow(), nullable=False)
-    count = db.Column(db.Integer, nullable=False)
-
-    def __init__(self, ip=None, port=None, measured_at=arrow.utcnow(), count=None):
-        self.ip = ip
-        self.port = port
-        self.measured_at = measured_at
-        self.count = count
-
-    @property
-    def ip(self):
-        if self._ip:
-            return long2ip(self._ip)
-
-    @ip.setter
-    def ip(self, value):
-        if value:
-            self._ip = ip2long(value)
-
-    def __repr__(self):
-        return '<ServerPlayerCount> {}:{}'.format(self.ip, self.port)
-
-
-# -----------------------------------------------------------
-# CLI commands
-
-
-@app.cli.command()
-def create_database():
-    """Delete then create all the database tables."""
-    db.drop_all()
-    db.create_all()
-
-
-@app.cli.command()
-def get_servers_player_count():
-    """Store the number of players on each servers."""
-    app.logger.info('Getting servers list')
-
-    scraper = rwr.DataScraper()
-
-    servers = scraper.get_servers()
-
-    for server in servers:
-        app.logger.info('  {} ({}, {})'.format(server.name, server.players.current, server.ip_and_port))
-
-        server_player_count = ServerPlayerCount()
-        server_player_count.ip = server.ip
-        server_player_count.port = server.port
-        server_player_count.count = server.players.current
-
-        db.session.add(server_player_count)
-
-    app.logger.info('Persisting data')
-
-    db.session.commit()
-
-    app.logger.info('Done')
-
-
-@app.cli.command()
-def delete_old_servers_player_count():
-    """Delete old servers player count."""
-    app.logger.info('Deleting old data')
-
-    # TODO Delete all data older than one month (exclusive)
-
-    app.logger.info('Done')
-
-
-@app.cli.command()
-@click.option('--gamedir', '-g', help='Game root directory')
-def extract_ranks_images(gamedir):
-    """Extract ranks images from RWR."""
-    context = click.get_current_context()
-
-    if not gamedir:
-        click.echo(extract_ranks_images.get_help(context))
-        context.exit()
-
-    app.logger.info('Extraction started')
-
-    extractor = rwr.RanksImageExtractor(gamedir, app.config['RANKS_IMAGES_DIR'])
-    extractor.extract()
-
-    app.logger.info('Done')
-
-
-@app.cli.command()
-@click.option('--gamedir', '-g', help='Game root directory')
-def extract_minimaps(gamedir):
-    """Extract minimaps from RWR."""
-    context = click.get_current_context()
-
-    if not gamedir:
-        click.echo(extract_minimaps.get_help(context))
-        context.exit()
-
-    app.logger.info('Extraction started')
-
-    extractor = rwr.MinimapsImageExtractor(gamedir, app.config['MINIMAPS_IMAGES_DIR'])
-    extractor.extract()
-
-    app.logger.info('Done')
+import routes
+import models
+import commands
 
 
 # -----------------------------------------------------------
